Added env variables

+++ movements.sh
@@ -1,8 +1,28 @@
+# variables that will be needed throughou the program
+configs_path=~/.configs/nvim
+profile_path=~/.profile
+
+# set environment variables
+while read -r line; do
+	env_split=($line)
+
+	cmd="export ${env_split[0]}=${env_split[1]}"
+	grep_res=$(grep -c "$cmd" $profile_path)
+
+	if [ $grep_res = 0 ]; then
+		echo $cmd >> $profile_path
+	fi
+done < .env
+
 # Copy vimrc
-cp ./.vimrc ~/.vimrc
+if [ ! -f ~/.vimrc ]; then
+	cp ./.vimrc ~/.vimrc
+fi
 
 # Copy bash profile
-cp ./.bash_profile ~/
+if [ ! -f ~/.bash_profile ]; then
+	cat ./.bash_profile >> $profile_path
+fi
 
 #### NeoVim
 # Installing packer
@@ -13,8 +33,10 @@ if [ ! -d $packer_path ]; then
 fi
 
 # Moving nvim directory
-configs_path=~/.configs
-if [ ! -d $configs_path ]; then
-	mkdir $configs_path
+if [ ! -d ~/.configs ]; then
+	mkdir ~/.configs
 fi
-cp -r ./nvim ${configs_path}/nvim
+
+if [ ! -d $configs_path ]; then
+	cp -r ./nvim ${configs_path}/nvim
+fi
